feat: introduce a renderer-factory that can produce a renderer
